+++ android/app/src/main/java/com/wooseokshim/babydiary/MainActivity.java
@@ -1,5 +1,0 @@
-package com.wooseokshim.babydiary;
-
-import com.getcapacitor.BridgeActivity;
-
-public class MainActivity extends BridgeActivity {}